hardening: prepared statements, PHP 7.4 idioms, and security fixes (#365)
* refactor(php74): use null coalescing (??) and ??= operators


* test: expand security test coverage for hardening changes

Add targeted tests for prepared statement migration, output escaping,
auth guard presence, CSRF token validation, redirect safety, and
PHP 7.4 compatibility. Tests use source-scan patterns that verify
security invariants without requiring the Cacti database.


* fix(js): migrate deprecated jQuery shorthand events to .on()/.off()


* style: align plugin with Cacti PHP coding standard

* ci: remove unavailable versioned Apache PHP package

* fix: pass integer timeout to process registration

* test: fix PHP regex quoting in security tests

* style: format security test fixtures

---------

+++ poller_intropage.php
@@ -95,7 +95,7 @@ intropage_debug('Intropage Starting Checks');
 
 // silently end if the registered process is still running, or process table missing
 if (function_exists('register_process_start')) {
-	if (!register_process_start('intropage', 'master', $config['poller_id'], read_config_option('intropage_timeout'))) {
+	if (!register_process_start('intropage', 'master', $config['poller_id'], (int) read_config_option('intropage_timeout'))) {
 		intropage_debug('Another Intropage Process Still Running');
 		exit(0);
 	}
